build scopes for organized db query code

+++ app/Http/Controllers/VoucherController.php
@@ -26,11 +26,9 @@ class VoucherController extends Controller {
                 ->withCount('voucher_records')
                 ->paginate(15)->withQueryString();
         } else {
-            // $vouchers = Auth::user()->vouchers()->whereDate('created_at', Carbon::today())->get();
-            // $vouchers = Voucher::where("user_id", Auth::id())->whereDate('created_at', Carbon::today())->get();
-            $vouchers = Auth::user()->vouchers()->whereDate('created_at', Carbon::today())
+            $vouchers = Auth::user()->vouchers()->today()
                 ->withCount('voucher_records')
-                ->get();
+                ->paginate(15)->withQueryString();
         }
         // return new VoucherCollectionResource($vouchers);
         return new RecentSaleOverviewResource($vouchers);
@@ -167,22 +165,14 @@ class VoucherController extends Controller {
 
     public function showTrash()
     {
-        if (Gate::allows('isAdmin')) {
-            $trashed_vouchers = Voucher::onlyTrashed()
-            ->withCount('voucher_records')
-            ->get();
-        } else {
-            $trashed_vouchers = Auth::user()->vouchers()->onlyTrashed()
-            ->withCount('voucher_records')
-            ->get();
-        }
-
+        $trashed_vouchers = Voucher::onlyTrashed()->ownByUser()->withCount('voucher_records')->get();
         return VoucherResource::collection($trashed_vouchers);
     }
 
-    // isAdmin middleware
     public function forceDelete(string $id)
     {
+        Gate::authorize('isAdmin');
+
         $voucher = Voucher::onlyTrashed()->find($id);
         if (is_null($voucher)) {
             return response()->json([
@@ -197,9 +187,10 @@ class VoucherController extends Controller {
         ], 204);
     }
 
-    // isAdmin middleware
     public function emptyBin()
     {
+        Gate::authorize('isAdmin');
+
         Voucher::onlyTrashed()->forceDelete();
         return response()->json([
             'message' => 'Trash has been emptied'
@@ -207,11 +198,11 @@ class VoucherController extends Controller {
 
     }
 
-    // isAdmin middleware
     public function recycleBin()
     {
-        Voucher::onlyTrashed()->restore();
+        Gate::authorize('isAdmin');
 
+        Voucher::onlyTrashed()->restore();
         return response()->json([
             'message' => 'Trash has been restored'
         ], 201);
